pyproject.toml
[tool.poetry]
name = "tcc-pos-graduacao-etl-duckdb"
version = "0.1.0"
description = ""
authors = ["gabriel-garciae <eng.gabrielgevangelista@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
duckdb = "^1.0.0"
pandas = "^2.2.2"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
